ulna/pyproject.toml
[project]
name = "ulna"
description = "A simple build system for Remige"
version = "0.1.0"
keywords = ["build system", "c"]

readme = "README.md"

requires-python = ">=3.12"
dependencies = [
    "anstrip>=0.2,<0.3",
    "option-python==0.0.2",
    "result>=0.17,<0.18",
]

[project.license]
file = "LICENSE"

[[project.authors]]
name = "Qexat"
email = "contact@qexat.com"

[project.urls]
repository = "https://github.com/qexat/remige/tree/main/ulna"

[project.scripts]
ulna = "ulna.__main__:main"

[build-system]
requires = ["setuptools >= 75.0"]
build-backend = "setuptools.build_meta"
